chore: app/layout 하위 컴포넌트 components/layout으로 위치 이동 (#190)
