+++ src/app/layout.tsx
@@ -2,9 +2,9 @@ import '../styles/globals.css';
 import { ThemeProvider } from '@/components/common/theme/ThemeProvider';
 import type { Metadata } from 'next';
 import localFont from 'next/font/local';
-import GlobalHeaderWrapper from './layout/header/GlobalHeaderWrapper';
-import GlobalFooter from './layout/footer/GlobalFooter';
-import ModalPortal from './layout/modal/ModalPotal';
+import GlobalHeaderWrapper from '../components/layout/header/GlobalHeaderWrapper';
+import GlobalFooter from '../components/layout/footer/GlobalFooter';
+import ModalPortal from '../components/layout/modal/ModalPotal';
 import ReactQueryProviders from '@/hooks/useReactQuery';
 import { Toaster } from '@/components/ui/toaster';
 import { cn } from '@/lib/utils';
