+++ src/components/layout/footer/GlobalFooter.tsx
